iteration10/requirements.txt
dearpygui==1.9.1
fastapi==0.103.0
plyer==2.1.0
pyserial==3.5
requests==2.28.2
uvicorn==0.23.2
websocket-client==1.6.2
